NF: move ImportReplace where it's needed
