+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionTask.java
@@ -31,6 +31,7 @@ import com.ichi2.anki.BackupManager;
 import com.ichi2.anki.CardBrowser;
 import com.ichi2.anki.CardUtils;
 import com.ichi2.anki.CollectionHelper;
+import com.ichi2.anki.DeckPicker;
 import com.ichi2.anki.R;
 import com.ichi2.anki.StudyOptionsFragment;
 import com.ichi2.anki.TemporaryModel;
@@ -1286,158 +1287,6 @@ public class CollectionTask<Progress, Result> extends BaseAsyncTask<Void, Progre
                 return new Triple(null, true, e.getMessage());
             }
             return new Triple<>(imp, false, null);
-        }
-    }
-
-
-    public static class ImportReplace implements TaskDelegate<String, Computation<?>> {
-        private final String mPath;
-
-
-        public ImportReplace(String path) {
-            this.mPath = path;
-        }
-
-
-        public Computation<?> task(@NonNull Collection col, @NonNull ProgressSenderAndCancelListener<String> collectionTask) {
-            Timber.d("doInBackgroundImportReplace");
-            Resources res = AnkiDroidApp.getInstance().getBaseContext().getResources();
-            Context context = col.getContext();
-
-            // extract the deck from the zip file
-            String colPath = CollectionHelper.getCollectionPath(context);
-            File dir = new File(new File(colPath).getParentFile(), "tmpzip");
-            if (dir.exists()) {
-                BackupManager.removeDir(dir);
-            }
-
-            // from anki2.py
-            String colname = "collection.anki21";
-            ZipFile zip;
-            try {
-                zip = new ZipFile(new File(mPath));
-            } catch (IOException e) {
-                Timber.e(e, "doInBackgroundImportReplace - Error while unzipping");
-                AnkiDroidApp.sendExceptionReport(e, "doInBackgroundImportReplace0");
-                return ERR;
-            }
-            try {
-                // v2 scheduler?
-                if (zip.getEntry(colname) == null) {
-                    colname = CollectionHelper.COLLECTION_FILENAME;
-                }
-                Utils.unzipFiles(zip, dir.getAbsolutePath(), new String[] {colname, "media"}, null);
-            } catch (IOException e) {
-                AnkiDroidApp.sendExceptionReport(e, "doInBackgroundImportReplace - unzip");
-                return ERR;
-            }
-            String colFile = new File(dir, colname).getAbsolutePath();
-            if (!(new File(colFile)).exists()) {
-                return ERR;
-            }
-
-            Collection tmpCol = null;
-            try {
-                tmpCol = Storage.Collection(context, colFile);
-                if (!tmpCol.validCollection()) {
-                    tmpCol.close();
-                    return ERR;
-                }
-            } catch (Exception e) {
-                Timber.e("Error opening new collection file... probably it's invalid");
-                try {
-                    tmpCol.close();
-                } catch (Exception e2) {
-                    Timber.w(e2);
-                    // do nothing
-                }
-                AnkiDroidApp.sendExceptionReport(e, "doInBackgroundImportReplace - open col");
-                return ERR;
-            } finally {
-                if (tmpCol != null) {
-                    tmpCol.close();
-                }
-            }
-
-            collectionTask.doProgress(res.getString(R.string.importing_collection));
-
-            try {
-                CollectionHelper.getInstance().getCol(context);
-                // unload collection and trigger a backup
-                Time time = CollectionHelper.getInstance().getTimeSafe(context);
-                CollectionHelper.getInstance().closeCollection(true, "Importing new collection");
-                CollectionHelper.getInstance().lockCollection();
-                BackupManager.performBackupInBackground(colPath, true, time);
-            } catch (Exception e) {
-                Timber.w(e);
-            }
-            // overwrite collection
-            File f = new File(colFile);
-            if (!f.renameTo(new File(colPath))) {
-                // Exit early if this didn't work
-                return ERR;
-            }
-            int addedCount = -1;
-            try {
-                CollectionHelper.getInstance().unlockCollection();
-
-                // because users don't have a backup of media, it's safer to import new
-                // data and rely on them running a media db check to get rid of any
-                // unwanted media. in the future we might also want to duplicate this step
-                // import media
-                HashMap<String, String> nameToNum = new HashMap<>();
-                HashMap<String, String> numToName = new HashMap<>();
-                File mediaMapFile = new File(dir.getAbsolutePath(), "media");
-                if (mediaMapFile.exists()) {
-                    try(JsonParser jp = AnkiSerialization.getFactory().createParser(mediaMapFile)) {
-                        String name;
-                        String num;
-                        if (jp.nextToken() != JsonToken.START_OBJECT) {
-                            throw new IllegalStateException("Expected content to be an object");
-                        }
-                        while (jp.nextToken() != JsonToken.END_OBJECT) {
-                            num = jp.currentName();
-                            name = jp.nextTextValue();
-                            nameToNum.put(name, num);
-                            numToName.put(num, name);
-                        }
-                    }
-                }
-                String mediaDir = Media.getCollectionMediaPath(colPath);
-                int total = nameToNum.size();
-                int i = 0;
-                for (Map.Entry<String, String> entry : nameToNum.entrySet()) {
-                    String file = entry.getKey();
-                    String c = entry.getValue();
-                    File of = new File(mediaDir, file);
-                    if (!of.exists()) {
-                        Utils.unzipFiles(zip, mediaDir, new String[] {c}, numToName);
-                    }
-                    ++i;
-                    collectionTask.doProgress(res.getString(R.string.import_media_count, (i + 1) * 100 / total));
-                }
-                zip.close();
-                // delete tmp dir
-                BackupManager.removeDir(dir);
-                return OK;
-            } catch (RuntimeException e) {
-                Timber.e(e, "doInBackgroundImportReplace - RuntimeException");
-                AnkiDroidApp.sendExceptionReport(e, "doInBackgroundImportReplace1");
-                return ERR;
-            } catch (FileNotFoundException e) {
-                Timber.e(e, "doInBackgroundImportReplace - FileNotFoundException");
-                AnkiDroidApp.sendExceptionReport(e, "doInBackgroundImportReplace2");
-                return ERR;
-            } catch (IOException e) {
-                Timber.e(e, "doInBackgroundImportReplace - IOException");
-                AnkiDroidApp.sendExceptionReport(e, "doInBackgroundImportReplace3");
-                return ERR;
-            }
-        }
-
-        @Override
-        public boolean requiresOpenCollection() {
-            return false;
         }
     }
 }
